requirements.txt
Django>=6.0
djangorestframework
djangorestframework-simplejwt
